Initial AuthService Refactor and ValidateAccessToken EndPoint
